feat: split client.ts into domain modules and add date-fns-tz
- Split 5153-line client.ts into 23 domain-specific modules + _core.ts
- Create barrel index.ts re-export for zero-change consumer imports
- Install date-fns-tz v3.2.0 and update getLocalDateStr() to use it
- Clean up unused imports across split modules

Domain modules: appointments, users, services, reviews, prescriptions,
invoices, clinical, mutations, dental, pharmacy, clinic, analytics,
booking, pharmacy-transactions, subscription, lab, parapharmacy,
radiology, medical-equipment, pediatrician, gynecologist,
ophthalmologist, kpis
